CHANNELMGR-1618 Put contextPayload in hints
The ChannelContentResource exposes an API for CRUD operations on
documents. The Resource delegates to a ContentService. The
implementation of the service uses the HintsInspector to decide if a
CRUD operation is allowed or not. The hints passed to the inspector
currently only contain the workflow hints. This is sometimes not
sufficient. The the context payload from the CmsSessionContext can
contain additional parameters that could be inspected by the hints
inspector to decide if a CRUD operation is allowed or not. So, before
passing the hints to the HintsInspector the ContentService must merge
all the 'hints' from the context payload with the workflow hints.

- Change the ContentService method args so that the contextPayload
  can be passed from the REST resource to the ContentService.
- EditingUtils should only contain static utility methods related to
  document CRUD operations. So move all the hints inspection code to
  the ContentServiceImpl.