CMS-10998 Support for a service through which we can post jcr node events
CMS wicket models at this moment are invalidated if for example it is a
JcrNodeModel and the backing jcr node gets modified. Whether this jcr
Node gets modified by the current cms visitor or via someone else,
every editor in the cms that has a JcrNodeModel for that node, gets its
model invalidated.
Now, there can be use cases where the model needs to be reloaded while
there is no actual jcr node change. For example when the workflow
integrates with an external BPM engine : If the change triggered by a
workflow step should alter the available workflow menu items, but the
workflow step only triggers a change in the bpm system, the menu won't reload.
Hence, we need a service to which we can post node events
(that will trigger a wicket model reload in the cms)