REPO-1910 Make sure the LockResourceImpl is thread-safe
Since #close can be invoked concurrently, make the #close method synchronized

Note that #getLock returns the MutableLock object *but* we need this object
because in other case instance equality is checked. This means that code using the
LockResource could cast the #getLock to MutableLock. Therefor I made the MutableLock
also thread-safe

Note that the #getHolder returns the Thread object that created the Lock. I
doubt whether Thread objects themselves are thread-safe (hard to google)
but I think this limitation might be acceptable. Not sure if we should
add to the javadoc that the returned #getHolder is not thread-safe?

Either way, most importantly, #close is now thread-safe since I have a use
case that could invoke #close from a stopping cluster node concurrently
with normal code execution (obviously chances are really small and impact
of incorrect concurrent access also small: Only an incorrect warning log)