HSTTWO-4183 Enable saving of a sub-set of component parameters
Without a Move-To header only existing component parameters are updated.
In that case the parameters of the variant are not removed first
anymore, so it becomes possible to supply only a sub-set of the
parameters and only update those.

It is save to slightly change the update-only logic because the Channel
Manager does not use it anymore: it always adds a Move-To header when
saving parameters, even when the variantId remains the same.