CHANNELMGR-1642 Rename OTHER_PROJECT to PART_OF_PROJECT
There are situations where a document cannot be edited while the active
and associated project for a document are the same. In that case we
want to show the same message as when the document belongs to another
project than the active one. This is also more in line with what is
shown in the content perspective.