CMS-11022 Move proxy functionality from resourceServlet to filter
The ResourceServlet functionality had to be extended to forward
baseUrl+path to baseUrl + path + index.html to proxy an angular 4
app running on its own location, e.g. http://localhost:9092 .

This could result in a potential vulnerability. Looking at this more
closely the proxy functionality does not belong in the resourceservlet.

For these reasons the proxy functionality was moved to a separate
filter that every project can add to their setup for local development.

It should only be used on local development, however if no proxies
are specified the filter will just call the next filter in the filter
 chain.