HSTTWO-4187 Support to inject a 'context augmenter' in DocumentsResource
In a downstream project, we need to be able to serve a link for a branch of a channel.
The downstream project has a customer channelManagerHstSiteProvider in
DelegatingHstSiteProvider, but we somehow need to be able to inject
knowledge into the context whether #getSite should return the master
or whether a branch should be tried.
via the generic DocumentContextAugmenter, we can store extra logic on
the HstRequestContext in a downstream project and use that later on
during the request processing