ESSENTIALS-1135 Make ContentBeansService(Impl) injectable
ContentBeansService implemented an intricate logic for accumulating
user feedback messages (as BeanWriterLogEntry) in a dedicated magic
field of the PluginContext, only to rewrite them immediately in the
BeanWriterUtils. I refactored ContentBeansService to directly and
consistently add user feedback messages to the passed-in UserFeedback
instance.
Also, the PluginContext no longer needs to be provided in the
constructor, which made ownership and control of the context's life-
cycle unnecessarily complicated. Instead, the caller owns the context
and passes it in through the API.

ContentBeansService still is an overly complex and under unit-tested
class. While the new interface describes the current API, the API is
still too ugly to freeze, which is why I keep it in the implementation
module for now.