ESSENTIALS-1135 Refactor JCR low-level access into JcrService
This change removes the dependency on the JCR-related methods of the
GlobalUtils by making them available through the injectable JcrService.
As many of the unit tests (sub-classes of BaseRepositoryTest) depend
on a custom access to the repository, this change involves some serious
refactoring in the testing infrastructure, to make sure there is only
a single instance (singleton) of the JcrService used throughout the
Spring application. In the unit-test context, the TestJcrService plays
that role.