ESSENTIALS-1135 Embed Vendor info in PluginDescriptor
The separate Vendor bean was is only used in the context of a
PluginDescriptor. It therefore seems cleaner to keep it as an inner
class. I also noticed that the unit test for the [de-]serialization of
PluginDescriptor lived in the dashboard, while the descriptor lives in
the API. I moved the unit test into the API module, rearranging some
Maven dependencies.