ESSENTIALS-1135 Add import of interpolated resource to JcrService
Importing a classpath XML file (after {{moustache}} interpolation) into
the JCR repository is a frequently-seen pattern in Essentials plugins.
By adding support for this pattern to the JcrService, the direct
dependency of a plugin on the TemplateUtils is removed, without having
to expose them through the API.

All in all, the Content Blocks plugin now no longer depends on the
plugin-sdk implementation module. Hurray!