HSTTWO-4199 Create a PingFilter next to the PingServlet
The PingServlet has as disadvantage that it needs to load the entire HST
model before it can return. This is because the PingServlet is after
the HstFilter. When this initial check takes very long, a load balancer
might send in the meantime many more ping requests.

This can cause startup problems
When using this PingFilter, directly a 200 is returned or a 503 (where
the latter means the HST webapp is still starting up). There are different
modes to check on (for example whether all hst config jcr nodes are loaded,
or the hst services is available, or just the repository)