ESSENTIALS-1137 Simplify aggregation of RebuildEvents
Essentials plugins can indicate that they require a project rebuild and
restart. They used to do so by constructing an event class, and posting
it on the eventBus. A listener would collect them into a list, from
which the dashboard collects them during each Ping.
The same functionality has been simplified inside the RebuildService,
which avoid the need for the event bus, collects the relevant information
(plugin IDs only) and provides them to the dashboard.
With this change, the last use of the API's PluginEvent is gone, and so
the PluginEvent interface is removed from the API.