ESSENTIALS-1138 Refactor blog importer configuration mechanism
The old code abused the instruction mechanism to persist the blog
importer configuration. This was both weird and hard tt understand /
maintain. I've replaced it by a REST Resource in combination with a
plain Java bean holding the deserialized configuration data.
Since the installation of the plugin now no longer attempts to access
the blog importer configuration in the repository, the initial rebuild
(to get the blog importer configuration in place by pulling in the
extra CMS dependency) is no longer necessary. I therefore refactored the
installation sequence to use the mavenDependency instruction, and the
plugin now installs with a single rebuild.