ESSENTIALS-1139 Refactor document wizard REST endpoint
The old code required extra encoding/decoding logic in both the front
end and the back end to massage the data into a generic PostPayload
format and extract it again to write it to the repository. This
refactoring introduces a plain Java bean for deserializing the
configuration data, and the dependency on the (overly) generic
PostPayload living in the SDK implementation module is gone.
As of this change, the document wizard plugin no longer depends on
the implementation module.