REPO-1915 Make sure to have latest jcr node states instead of possibly stale nodes
Since the lock via the LockManager does not involve any jcr locking, it
does *never* trigger a cluster sync. This means that we can get the lock
for a node that has been processed and finished just by another cluster
node that also already freed the lock again. Making sure that after we
obtained the lock to invoke a session refresh should result in that
#isPendingTrigger returns false if another cluster node already did
process the job.