ESSENTIALS-1137 More refactoring of PluginContext
With the PluginContext having lost much of its weight and importance,
it was now unused in many method signature and interface. This commit
focuses on removing this superfluous passing-around of PluginContext
instances.

This commit also cleans up the ProjectSettings API to only contain
getters, and only for fields potentially relevant to plugins. Dashboard-
internal data can be accessed through the ProjectSettingsBean, as can
the settings be adjusted.