ESSENTIALS-1137 Remove unnecessary complexity in PluginConfigService
PluginConfigService used to be an SDK API interface for (de-)serializing
Documents to/from XML files. This mechanism is intended for the
Dashboard only, and used in the plugin InstallStateMachine and the
SettingsServiceImpl (for managing the project-global settings).

This refactoring removes these concepts from the SDK API (not relevant
there) and removes unnecessary abstraction layers.