ESSENTIALS-1137 Remove usage of instructions from REST and Gallery tools
Instead of "manually" executing an instruction set, I refactored the
FileInstruction to use the ProjectService for copying files to the
project or deleting files from the project. The REST resource if the
REST plugin can now use the ProjectService (and JcrService) directly,
rather than taking the elaborate detour through the instruction
mechanism, which should really be dashboard-internal.