CHANNELMGR-1660 Show content editor again
The content editor has become a separate component so it can easily be
reused. It contains the shared space toolbar, the error feedback and the
editable document fields. The styling of the the content editor and its
fields has been split from the styling of the right side panel.

The content editor does not contain the form: that has moved to the
editContentMain class so it can easily be used to disable the save
button when the form is invalid.

The content editor shows the document loaded by the content service.
The content service also takes care of saving the document when the
edit-content state is exited. It therefore needs to now whether the
edited document is dirty, which components have to report explicitly
by calling markDocumentDirty(). The primitive field and CKEditor field
now do this in their onChange handlers.

To get the correct scrolling behavior (i.e. only the editor scrolls and
the bottom toolbar not) the main and tools part of edit-content are no
longer a component but a separate controller and template. If they are
components, ui-router renders an additional tag for the main component
that does not have the necessary flexbox styling.