ESSENTIALS-1137 Clean up installation instruction infrastructure
Many instruction type unit tests used an executor and a wrapping
instruction-set, this is unnecessary complexity in the tests, I removed
them. Furthermore, the instruction package and associated classes
contained a bunch of unused code, which I also removed. And I made the
naming of the CommonsInstructionPackage more comprehensible by renaming
it to SkeletonInstrictionPackage.