CHANNELMGR-1660 Don't use URL for hippo-cm.channel state params
When the channelId is part of the URL, a reload of the app (e.g. because
Webpack reloads the app because of a source code change) will directly
navigate to the hippo-cm.channel state. However, the initialization code
currently expects that the CMS passes a complete 'channel' object that
also includes a mountId, and it also expects a projectId. That
initialization logic should be refactored to be able to use proper URL
state. For now we keep the old logic.