CHANNELMGR-1660 Move form and save/close buttons into contentEditor
It's cleaner to make the content editor own the form. It can then watch
the form's dirty state and update the dirty state in the content editor
service in a central location instead of having each leaf form component
take care of that itself.

When the form is part of the editor but that buttons are not, the editor
would have to expose the $valid state (to enable the save button). But
the outside should also be able to set the form to pristine again after
saving. That combination is hard, so better move the button into the
component. The outside only provides a label for the 'cancel' state, and
callbacks for when the document is saved or canceled.