REPO-1907 Get action from args when method is triggerAction
We introduced DocumentWorkflow#triggerAction() and implemented it in the
DocumentWorkflowImpl.As a consequence when this new method is used the
WorkflowInvocationHandler invokes the triggerAction method while the
actual action is contained in the args array. After invocation of the
method an event is published. The published event does however not
contain the action that was actually performed, but only the method.

I fixed this by explicity checking the method name for "triggerAction"
and then finding the corresponding Action in the args. If an action
exists it will be set on the event.