ESSENTIALS-1137 Get rid of Restful marker interface
Many serializable bean classes in Essentials (and plugins) implemented
the empty Restful marker interface. Its primary usage was the
RestfulList, which encapsulates a list of various such Restful beans.
In reality, the list was only used for homogenous lists of type
PluginDescriptor, MessageRestful or KeyValueRestful. These usages have
been refactored, and RestfulList is now only used in combination with
PluginDescriptor, when parsing/deserializing a list of them from JSON.
The Restful marker interface is gone, and the remaining REST-scoped
classes in the SDK API are all pure beans now.