ESSENTIALS-1137 Refactor PluginContext into PlaceholderService
The content of PluginContext has already gradually been brought down to
only wrap a map of placeholder objects, to be used when working with
template resources during plugin installation or configuration.

To make the intent of this functionality much clearer, It is now
provided through the PlaceholderService of the SDK API. The creation of
the placeholder data has been pushed as close to its usage as possible,
such that it is more comprehensible what kind of data is carried inside
that map. This change caused the removal of the PluginContext and its
factory service, as well as a change to the Instruction SPI.
The installation logic has been simplified a little, for better
maintainability. The handling of installation parameters is now more
explicit, and clearly separated from the placeholder data management.
Supported placeholder names who show up in the plugin's template
resources are now part of the API.