ESSENTIALS-1137 Make plugin-used dashboard rest endpoints explicit API
Two endpoints of the dashboard's DocumentResource used to return a list
of generic KeyValueRestful instances. The plugin would then guess the
meaning of the 'key' and the 'value', and the implementation would be
bound to these two fields. To avoid the guessing and make the API more
future-proof, I introduce the ContentTypeInstance and TemplateQuery
beans to the SDK API, which are now returned by the corresponding REST
endpoints, and consumed by the Selections and Document Wizard plugins.
KeyValueRestful is no longer used and has been removed.