CHANNELMGR-1539 Update component parameter after step 2
The logic to save a component parameter moved to the new
HstComponentService so it can be reused in the OverlayService and the
Step2Service. The logic to pick a path also moved to the
HstComponentService since it seems closely related, and reduces the
(huge) size of the OverlayService a little bit. Both pickPath and
setPathParameter accept a basePath; for relative paths that will be the
channel content root, otherwise it will be empty. This way the fetching
of the channel content root can be done outside the HstComponentService.

The overlay passes the necessary information about the component
parameter to the CreateContentService. Step 1 only remembers it
so it can be passed to step 2.

Saving the document after step 2 triggers a notification "Document
saved". If a component parameter is provided the path to the document is
then stored in the parameter. This triggers a second notification, or an
error when it fails (which can happen if the component is locked by
someone else concurrently).

The OverlayService now also checks for buttons with a componentParameter
but without an enclosing container item. Those indicate a developer
error. A warning is logged and the component parameter is ignored.

Still todo: add more unit tests.