ESSENTIALS-1151 Implement and unit-test new installation state machine
The new state machine knows two more states: "boarding pending" and
"installation pending". These indicate that a plugin is waiting to
board or install, and cannot do so because a dependent plugin hasn't
reached the proper installation state yet. These dependencies have been
added to the PluginDescriptor: a plugin can depend on multiple other
plugins, and for each dependent plugin, one can specify in what minimal
state it should be before the requested plugin can board or be
installed.

The front-end and the plugin descriptors have not yet been changed.