CHANNELMGR-1539 Don't save component parameter when it's not specified
The save-component-parameter-after-creating-content logic used to be
"if we're located inside a component, update the component parameter".
That wrong, since a 'create content' button can also be placed inside a
component yet without a component parameter. The typical example is a
generic "+" button to create new content inside a query-based component,
e.g. a news list.

The fix is to check whether there's a component parameter specified, and
only then update that parameter.