HSTTWO-4222 Use parameter for active render variant in manageContent tag
The active render variant of a component can be set in the
HstComponentWindow attribute ContainerConstants.RENDER_VARIANT.
By default it's not set.

The manage content tag uses the active render variant to determine which
component parameter value to use. When no render variant is set, or
when it's empty, the default (non-prefixed) parameter value is used.
When there is a specific render variant set, the parameter value
prefixed with that render variant is used.