ESSENTIALS-1137 Move ApplicationModule into test sources
ApplicationModule has been trimmed down during the SDK API refactoring,
and all that's left is the component scan configuration duplicated in
the dashboard's applicationContext.xml. Analysing where/how it's still
used, I noted that it is only relevant to the (SDK implementation( unit
tests, so I moved it there to express that fact more appropriately.