CHANNELMGR-1515 Use correct render variant to update component parameter
The code used to save the component parameter for the render variant of
the whole page. That's not correct: a component may not have a variant
for the render variant of the page.

The render variant of a component is now available as meta-data of a
component in the property HST-Render-Variant. When not present or
an empty string, the default variant is used ('hippo-default').

Since a variant can contain fancy characters, it has to be URI-encoded
before being used in the path of a REST call. Fixing this generically in
the HST service (i.e. always URI-encoding all path parameters) is tricky
since those path elements are corrected for leading and trailing slashes
and leading and trailing spaces in those parameters. Those would be
wrongly URI-encoded. Only URI-encoding the variant seemed safer.

The first six manage-content buttons in the overlay fixture were located
in a container instead of in a component. That's conceptually wrong,
since those buttons will always be located inside a component or not
inside anything, but never directly inside a container. It also broke
the code, since a container did not have a getRenderVariant method.
So those buttons moved inside 'component B' (which did not break any
tests).

One manage content button has been added to the overlay fixture to
capture the pick-path logic for a button inside a component with a
non-default render variant. It's located inside the
'component-with-experiment', which now got a different rendering variant
in its meta-data. To keep the numbered manage-content button scenarios
in order, this new button got scenario number 7 and the numbers of the
remaining scenario's increased by one (both in the fixture and in the
tests).