ESSENTIALS-1151 Create angular projectService
This gets rid of many of the remaining dirty $rootScope usages.
projectService caches the project settings, so the new solution is also
performing better.
I improved the pluginService by moving the sorting and icon-selection
into the getChangeMessages method, and I fixed the "simple install
plugin" to get back a working uib-accordion for showing the parameter-
dependent change messages.