CHANNELMGR-1520 Recognize compound fields explicitly
Instead of assuming that everything that's not a choice field is
a compound, we now explicit recognize compounds fields: one of their
required supertypes is always "hippo:compound". That prevents the
code from recognizing compound-like types like hippogallerypicker:imagelink
as a compound. These will now be parsed an UNKNOWN types instead,
which makes it possible to later recognize them as unknown required
types.