CHANNELMGR-1520 Fix reporting of unsupported field types
They were not included anymore because FieldTypeUtils#determineFieldType
returned "Unknown" instead of an empty string for unsupported fields.

Don't include the 'unsupportedFieldTypes' property in the JSON output
unless it's not empty, otherwise the frontend will erroneously report
a usage statistic about it with an empty array of type names (an empty
array is truthy in JavaScript).