CHANNELMGR-1696 not set a default document location when no root path is defined
When the manageContent tag does not set a root path, the location field must not be filled with the channel root path by default. The user must pick a folder using the picker. The picker needs a path to be able to work though, so when no root path is given, a default picker path is used, which is the channel root path.

The backend document service needs root path and default path values to be able to create the document. The default path is used to optionally create folders that not yet exist. In case of no defined root path, the user must always pick a path which then always exist. In this case the picker must be able to set the root path in the step1 service, and the default path should be empty.