CHANNELMGR-1520 Process unknown fields recursively
Renamed FieldsInformation#allRequiredFieldsIncluded to
FieldsInformation#canCreateAllRequiredFields. A required choice field is
included/supported, but still cannot be created since there's no way yet
to add one of the choices in the UI.

Unknown fields in compounds (and compounds in compounds, etc.) are now
also noted.

Removed the UnknownFieldType again. It was only used for processing the
FieldsInformation, which can also be done without creating a FieldType
class via reflection.