CMS-11070 Use action for method if present
With projects we override DocumentWorkflow#triggerAction(). Result is
that the  hippolog:methodName in the event is always "triggerAction"
and the hippolog:action is the action that is actually performed.

Unfortunately the EventModel currently uses the hippolog:methodName
as part of the lookup key for a translation, which is always
"triggerAction" for project document workflows. To fix this, we
should only use the methodName if the action is absent.