CHANNELMGR-1696 set the root path to the channel content path when it is missing but there is a default path
Developers can configure the manageContent tag without a rootPath but with a defaultPath. This suggests a location for new documents within a channel root, but it will work within any channel. In those cases the document location field is populated with the channel content path as root path plus the default path.