CHANNELMGR-1702 Use channelId of parent if channel has a branchId
Currently the channel store only contains master channels. So the Ext
javascript code in the client only knows about master channels too.
If a channel branch must be opened in the channel editor we can't use
the channelId of the branch, but instead use the channelId of it's
parent, i.e. use channelDocument#branchOf(). Then we must also tell the
javascript code the id of the branch so that the hst can be requested to
set the active branch (via a rest call) for rendering the preview of
the site.

Preferably we would of course want to populate the channel store with
all channels (master and branches), which would make it possible to just
use the channelId directly without the branchId. We can't do that
however because then the channel overview would also show all the
branches as channels. The user would not be able to tell which channel
is master and which are branches because the look the same. To implement
this properly we first need a good design.