REPO-1934 Make sure PersistedEventListeners can always proceed with events
Before this commit, the logic was such that first an attempt was
done to find which listener should be used for processing events.

This was done in BroadcastModule#getNextJob which would return
the listener (say A) with the oldest 'hippobroadcast:lastProcessed' value.

The problem however was that if all the events to process
would be filtered for listener A because of the #getCategory to
listen to (and no events for the category), the listener A
would not process any events and would not update the
'hippobroadcast:lastProcessed' value, resulting in the next
run again to use Listener A (while there might already be
more than enough events for Listener B which doesn't
filter on a category)

The changes in this commit first of all change
BroadcastModule#getNextJob to not return a single listener
to process events for, but instead *all* listeners that
are registered, where every listener has its own 'lastProcessed'
value and Classloader stored in a Triple. We cannot use a
global last processed unfortunately because of the
listener.onlyNewEvents() concept.

After we have all the listeners, a BroadcastJobImpl
containing all listeners and the globalLastProcessed. In
the Broadcaster.JobRunner we now fetch *all* events
after globalLastProcessed and then give every event
to BroadcastModule.BroadcastJobImpl#publish in which we check
for every listener whether it should process the event.

The big difference is that instead of querying for all
event items *per* listener, one listener after another,
we query once for all event items, and feed them to every
listener is the listener doesn't filter out the event