CHANNELMGR-1520 Move 'alternative step 2' logic outside content editor
The message shown above the content editor differs in state
'edit-content' and 'create-content-step-2'. Instead of having the
content editor know both messages, transclude the message and show it
based on a single input property 'showMessage'. This also removes the
need for the 'switchEditor' callback, since each state template
completely owns the message on top of the content editor and any
behavior attached to it.

Whether to allow saving now depends on an additional input parameter
'allowSave'. In 'edit-content' state saving is always allowed, but in
'create-content-step-2' it depends on whether all required fields in
the document type can be created.