CHANNELMGR-1520 Include field info of empty (invalid) compounds
A compound may be empty, which makes it invalid, so it is not included
in the document type. However, the compound may be empty because it
only contains unsupported fields. This fields information of such
a compound should therefore be added, otherwise those unsupported fields
are not reported nor included in the overall fields information. This
means that the fields information when initializing a field should
always be added, regardless whether the field is valid, since it may
contain information about other unsupported fields.