REPO-1939 Fix hippo lock table creation for MySQL
Make sure that *if* the Hippo lock table exists, that it really contains
an index on lockKey (regardless whether it is a primary key or a unique
index). If it doesn't contain one, the lock table was not created
correctly and contains locks that do not work. Hence, truncating the
lock table is ok. After that, correct the table scheme. Right after
the table scheme has been fixed, other running cluster nodes can log
some errors wrt locks. This is not a problem and expected