HSTTWO-4238 use channel root path as default for the picker root path
If the tag attribute rootPath is not available and the @JcrPath annotation of the component parameter is not given, use the channel root path as fallback/default for the picker root path. This prevents that the picker dialog allows the user to pick documents from outside the channel's content folder.

In case this IS actuallly wanted, then the rootPath attribute could be configured with something like '/content/documents' to allow for this.

Due to the fact that now in many more cases the pickerRootPath by default is provided, a number of unit tests had to be extended to cover this.