CHANNELMGR-1726 Fix picking a relative path for manage content parameter
The picker should always pick an absolute path. The HstComponentService
makes the path relative (to the given basePath) when the parameter is
relative.

ManageContentLink#isParameterValueRelativePath assumed that the property
was a boolean while it was in fact a string (like all other "boolean"
values in HTML comments). So we should compare it to the string 'true'.