HSTTWO-4244 Create preview configuration for branches if absent
PS or partners might be used to or have a script that removes -preview
configs during a new deployment, including preview configs of branches.
For master branches the -preview config is already created automatically
if it's missing, but for branch config not.

Solve this by also recreating -preview config of the branches whenever
master -preview config is recreated. The channel manager will call
startEdit for the master before it switches to a branch, so it's safe to
implement this in the MountResource#startEdit().