CHANNELMGR-1725 show an extra continue button for create content alternative step 2
Alternative step 2 occurs when a document type has required fields that are not supported by the editor yet. The user cannot save the document in the channel manager and must continue editing in the content perspective. There the document can be saved.

Two components use the content editor: the edit content component (for existing documents), and the create content step 2 component. With transclude they passed the content of the message at the top. The transclude is changed to support multiple slots: one for the message, one for extra buttons.

More of the message markup has moved to both parent component templates now. The multiple slot transclusion setup also allows (demands?) that the content-editor component gets names child component parts, which makes the transclusion better readable and less magic.