CHANNELMGR-1721 Focus original field when Ext picker dialog is cancelled
The ExtLinkPickerFactory now accepts an additional 'cancelled' callback
that is called when the opened dialog is cancelled (by clicking Cancel,
or when pressing Escape).

The ChannelEditor forward a 'cancelled' callback via a 'path-cancelled'
event to the document location field in the AngularJs app. The document
location field then focuses its input overlay. That div needs a tabindex
of -1 to become focusable. As a result, pressing Escape after closing
the Ext link picker dialog closes the right side-panel again (since the
keypress is then going to the focused document location field).

The ExtLinkPicker fields in the component properties dialog now use a
similar mechanism. They also listen to the cancelled callback, and
focus their internal 'renderTextField'. That also needs a tabindex of -1
to become focusable. As a result, pressing Escape after the picker
dialog is closed closes the component properties dialog.