CHANNELMGR-1731 Only show overlay if project in "Unapproved"
Only show the overlay if there is a selectedProject that has the
state "Unapproved".

A listener implementation is impossible, because the project service
gets loaded after the overlay service. So register to a a listener
in the project service is not going to work.

Alternatively calling the overlay service from the project service
creates a circular dependency.

Calling the setter is needed because the overlays have to be in sync
with the toggles.