CMS-10913 Add beforeShowDialog and use new DialogConfig
The DialogManager allows the frontend to provide a custom configuration
in the form of a JSON object. This configuration is merged with the
default configuration that is passed as an IPluginConfig in the
constructor.
To simplify the handling of other parameters coming from the frontend,
I introduced the hook #beforeShowDialog that is called with all request
parameters.

The DialogConfig replaces the ConfigProvider.